add sorting for galaxy queries, sort by mass * prob when looking for galaxies in skymaps, make it the default strategy for galaxy targeting with gwemopt, let users pick a galaxy catalog from the analysis section of the GcnPage
